ml/requirements.txt
requests
python-dotenv
pandas
pyarrow
tqdm
networkx
rapidfuzz
scikit-learn
google-cloud-bigquery
google-cloud-bigquery-storage
db-dtypes
